# tutorials/building-microservices-youtube/product-api/handlers/put.go
package handlers

import (
	"net/http"

	"github.com/LeeDark/go-microservices-starter/tutorials/building-microservices-youtube/product-api/data"
)

// swagger:route PUT /products products updateProduct
// Update a products details
//
// This will update a product in the data store
//
// Responses:
//   204: noContentResponse
//   404: errorResponse
//   422: errorValidation

// Update handles PUT requests to update products
func (p *Products) Update(rw http.ResponseWriter, r *http.Request) {
	rw.Header().Add("Content-Type", "application/json")

	// fetch the product from the context
	prod := r.Context().Value(KeyProduct{}).(data.Product)
	p.l.Debug("Updating record", "id", prod.ID)

	err := p.productsDB.UpdateProduct(prod)
	if err == data.ErrProductNotFound {
		p.l.Error("Product was not found", "error", err)

		rw.WriteHeader(http.StatusNotFound)
		data.ToJSON(&GenericError{Message: "Product not found in database"}, rw)
		return
	}

	// write the no content success header
	rw.WriteHeader(http.StatusNoContent)
}